Futher clean up the test codes
1. Move DnuListTestContext to its own file;
2. Remove unecessary methods;
3. Use Environment.NewLine to split output.
