+++ test/Microsoft.Framework.CommonTestUtils/DnuTestUtils.cs
@@ -11,55 +11,15 @@ namespace Microsoft.Framework.PackageManager
 {
     public static class DnuTestUtils
     {
-        public static int ExecDnu(string runtimeHomePath, 
-                                  string subcommand, 
-                                  string arguments, 
-                                  out string stdOut, 
-                                  out string stdErr,
-                                  IDictionary<string, string> environment = null, 
-                                  string workingDir = null)
-        {
-            string program, commandLine;
-            BuildProcessInformation(runtimeHomePath, subcommand, arguments, out program, out commandLine);
-
-            var exitCode = TestUtils.Exec(program, commandLine, out stdOut, out stdErr, environment, workingDir);
-
-            return exitCode;
-        }
-
-
         public static int ExecDnu(string runtimeHomePath,
                                   string subcommand,
                                   string arguments,
-                                  out string[] stdOut,
-                                  out string[] stdErr,
+                                  out string stdOut,
+                                  out string stdErr,
                                   IDictionary<string, string> environment = null,
                                   string workingDir = null)
         {
             string program, commandLine;
-            BuildProcessInformation(runtimeHomePath, subcommand, arguments, out program, out commandLine);
-
-            string output, error;
-            var exitCode = TestUtils.Exec(program, commandLine, out output, out error, environment, workingDir);
-
-            stdOut = output.Split('\n');
-            stdErr = error.Split('\n');
-
-            return exitCode;
-        }
-
-        public static int ExecDnu(string runtimeHomePath,
-                                  string subcommand,
-                                  string arguments,
-                                  IDictionary<string, string> environment = null, 
-                                  string workingDir = null)
-        {
-            string stdOut, stdErr;
-            return ExecDnu(runtimeHomePath, subcommand, arguments, out stdOut, out stdErr, environment, workingDir);
-        }
-
-        private static void BuildProcessInformation(string runtimeHomePath, string subcommand, string arguments, out string program, out string commandLine)
-        {
             var runtimeRoot = Directory.EnumerateDirectories(Path.Combine(runtimeHomePath, "runtimes"), Constants.RuntimeNamePrefix + "*").First();
             if (PlatformHelper.IsMono)
             {
@@ -72,6 +32,20 @@ namespace Microsoft.Framework.PackageManager
                 var dnuCmdPath = Path.Combine(runtimeRoot, "bin", "dnu.cmd");
                 commandLine = string.Format("/C {0} {1} {2}", dnuCmdPath, subcommand, arguments);
             }
+
+            var exitCode = TestUtils.Exec(program, commandLine, out stdOut, out stdErr, environment, workingDir);
+
+            return exitCode;
+        }
+
+        public static int ExecDnu(string runtimeHomePath,
+                                  string subcommand,
+                                  string arguments,
+                                  IDictionary<string, string> environment = null,
+                                  string workingDir = null)
+        {
+            string stdOut, stdErr;
+            return ExecDnu(runtimeHomePath, subcommand, arguments, out stdOut, out stdErr, environment, workingDir);
         }
     }
 }
